pyproject.toml
[project]
name = "robinjacobs-github-io"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.3",
    "pillow>=12.1.0",
    "python-multipart>=0.0.26",
    "uvicorn>=0.44.0",
]
